Adjusted build scripts to work with Travis

+++ build/tasks/install-spark-dev.coffee
@@ -3,6 +3,7 @@ fs = require 'fs-extra'
 request = require 'request'
 Decompress = require 'decompress'
 cp = require '../../script/utils/child-process-wrapper.js'
+_s = require 'underscore.string'
 
 workDir = null
 
@@ -12,43 +13,54 @@ module.exports = (grunt) ->
   grunt.registerTask 'install-spark-dev', 'Installs Spark Dev package', ->
     done = @async()
     workDir = grunt.config.get('workDir')
+    particleDevPath = path.join(workDir, 'node_modules', 'spark-dev')
+    particleDevVersion = grunt.config.get('particleDevVersion')
 
-    tarballUrl = 'https://github.com/spark/spark-dev/archive/v' + grunt.config.get('particleDevVersion') + '.tar.gz'
-    tarballPath = path.join(workDir, 'sparkdev.tar.gz')
-    sparkDevPath = path.join(workDir, 'node_modules', 'spark-dev')
-    r = request(tarballUrl)
-    r.on 'end', ->
-      decompress = new Decompress()
-      decompress.src tarballPath
-      decompress.dest sparkDevPath
-      decompress.use(Decompress.targz({ strip: 1 }))
-      decompress.run (error) ->
-        if error
-          throw error
+    installDependencies = (done) ->
+      # Build serialport
+      packageJson = path.join(workDir, 'package.json')
+      packages = JSON.parse(fs.readFileSync(packageJson))
+      process.chdir(particleDevPath);
+      env = process.env
+      env['ATOM_NODE_VERSION'] = packages.atomShellVersion
+      env['ATOM_HOME'] = if process.platform is 'win32' then process.env.USERPROFILE else process.env.HOME
+      options = {
+        env: env
+      }
 
-        fs.unlinkSync tarballPath
-        fs.removeSync path.join(sparkDevPath, 'build')
-        fs.removeSync path.join(sparkDevPath, 'docs')
+      if process.platform == 'win32'
+        command = '..\\..\\apm\\node_modules\\atom-package-manager\\bin\\apm.cmd'
+      else
+        command = '../../apm/node_modules/atom-package-manager/bin/apm'
 
-        # Build serialport
-        packageJson = path.join(workDir, 'package.json')
-        packages = JSON.parse(fs.readFileSync(packageJson))
-        process.chdir(sparkDevPath);
-        env = process.env
-        env['ATOM_NODE_VERSION'] = packages.atomShellVersion
-        env['ATOM_HOME'] = if process.platform is 'win32' then process.env.USERPROFILE else process.env.HOME
-        options = {
-          env: env
-        }
+      verbose = if !grunt.option('verbose') then '' else ' --verbose'
+      cp.safeExec command + ' install' + verbose, options, ->
+        injectPackage 'spark-dev', particleDevVersion
+        done()
 
-        if process.platform == 'win32'
-          command = '..\\..\\apm\\node_modules\\atom-package-manager\\bin\\apm.cmd'
-        else
-          command = '../../apm/node_modules/atom-package-manager/bin/apm'
+    if _s.endsWith(particleDevVersion, '-dev')
+      # Copy current sources
+      fs.copySync path.join(__dirname, '..', '..'), particleDevPath
+      installDependencies done
+    else
+      # Download the release
+      tarballUrl = 'https://github.com/spark/spark-dev/archive/v' + particleDevVersion + '.tar.gz'
+      tarballPath = path.join(workDir, 'sparkdev.tar.gz')
 
-        verbose = if !grunt.option('verbose') then '' else ' --verbose'
-        cp.safeExec command + ' install' + verbose, options, ->
-          injectPackage 'spark-dev', grunt.config.get('particleDevVersion')
-          done()
+      r = request(tarballUrl)
+      r.on 'end', ->
+        decompress = new Decompress()
+        decompress.src tarballPath
+        decompress.dest particleDevPath
+        decompress.use(Decompress.targz({ strip: 1 }))
+        decompress.run (error) ->
+          if error
+            throw error
 
-    r.pipe(fs.createWriteStream(tarballPath))
+          fs.unlinkSync tarballPath
+          fs.removeSync path.join(particleDevPath, 'build')
+          fs.removeSync path.join(particleDevPath, 'docs')
+
+          installDependencies done
+
+      r.pipe(fs.createWriteStream(tarballPath))
